+++ quantum/keymap_introspection.c
@@ -4,7 +4,7 @@
 // Pull the actual keymap code so that we can inspect stuff from it
 #include KEYMAP_C
 
-#include "keymap_wrapper.h"
+#include "keymap_introspection.h"
 
 #define NUM_KEYMAP_LAYERS ((uint8_t)(sizeof(keymaps) / ((MATRIX_ROWS) * (MATRIX_COLS) * sizeof(uint16_t))))
